src/etlbronze/pyproject.toml
[tool.poetry]
name = "etlbronze"
version = "0.1.0"
description = ""
authors = ["Kaylane Vasconcelos <kaylanedecassia10@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
prefect = "^3.1.2"
pandas = "^2.2.3"
psycopg2-binary = "^2.9.10"
flask = "^3.1.0"
flasgger = "^0.9.7.1"
python-dotenv = "^1.0.1"
boto3 = "^1.35.66"
clickhouse-connect = "^0.8.6"


[tool.poetry.group.dev.dependencies]
pytest-html = "^4.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"